// packages/web/src/generated/rust-types/SshFileEntry.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type SshFileEntry = { name: string, isDir: boolean, size: number, 
/**
 * Octal permission string, e.g. "0755"
 */
mode: string | null, 
/**
 * Owner user name
 */
owner: string | null, 
/**
 * Owner group name
 */
group: string | null, 
/**
 * ISO 8601 modified time
 */
modifiedAt: string | null, };